Blocksworld/requirements.txt
dash==2.8.1
networkx==2.6.3
openai==0.26.5
pandas==1.3.5
plotly==5.13.0
pygame==2.1.3
PyYAML==6.0
tarski==0.8.2
tqdm==4.64.1
